Provide the OpenGL vendor, version and visual information in the gfx list window command.
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/cmgui/trunk@2013 3c1deb5b-d424-0410-962d-aba41a686d42

+++ source/three_d_drawing/ThreeDDraw.c
@@ -2005,7 +2005,7 @@ Returns the visual ID actually used by this X3d widget.
 	if ((drawing_widget=(ThreeDDrawingWidget)widget)&&
 		(True==IsThreeDDrawing(widget)))
 	{
-		visual_id=(drawing_widget)->three_d_drawing.normal_buffer.visual_information->visualid;
+		visual_id=(int)(drawing_widget)->three_d_drawing.normal_buffer.visual_information->visualid;
 	}
 	else
 	{
